+++ lesson2_vera/1_vector_distance.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require_relative 'lib/vector'
+require_relative 'vector/vector'
 
 vector = Vector.new(3, 6, -1, 5)
 puts vector.distance
