+++ lesson2_vera/3_convert_file_size.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+require_relative 'float/float_byte_to'
+
+file_name = ARGV.first || 'gothic2.mp3'
+file_size = File.size(file_name).to_f
+puts "Размер файла : #{file_size.show_size}"
