+++ lesson2_vera/README.md
@@ -1,0 +1,2 @@
+2. Так как "завал" на работе , не когда было учитывать всю логику работы программы, потому проверку на присутствие сотрудника в списке  и т.д. не успела сделать
+3. Тут решила открыть не Integer, а Float. Мне больше нравится видеть более точные данные в ответе программы.
