+++ lesson2_vera/float/float_to_byte.rb
@@ -1,0 +1,20 @@
+# frozen_string_literal: true
+
+class Float
+  TO_BYTES = 2**10
+  def kilobytes
+    self * TO_BYTES
+  end
+
+  def megabytes
+    kilobytes * TO_BYTES
+  end
+
+  def gigabytes
+    megabytes * TO_BYTES
+  end
+
+  def terabytes
+    gigabytes * TO_BYTES
+  end
+end
