+++ lesson2_vera/unit/collective.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+COLLECTIVE = [
+  ['Иванов Иван Иванович', 'Team Leader'],
+  ['Петров Петр Петрович', 'Backend Developer'],
+  ['Владимиров Владимир Владимирович', 'Backend Developer'],
+  ['Олегов Олег Олегович', 'Frontend Developer'],
+  ['Аннушкина Анна Андреевна', 'Frontend Developer'],
+  ['Татианко Татьяна Тимуровна', 'QA Engeneer'],
+  ['Янейко Янина Ярославовна', 'Designer']
+].freeze
